+++ src/framefilter/BackgroundSubtractorMOG.h
@@ -64,7 +64,7 @@ private:
      * Configure the GPU to perform background subtraction.
      * @param index_ Index of the GPU to use for processing
      */
-    void configureGPU(int64_t index_);
+    void configureGPU(size_t index_);
 
     cv::Ptr<cv::cuda::BackgroundSubtractorMOG> background_subtractor_;
     cv::cuda::GpuMat current_frame_, background_mask_;
